Abstract get-badges API calls
